+++ apps/party/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gtmgrid/party
 
+## 0.13.0
+
+### Patch Changes
+
+- @gtmgrid/auth@0.13.0
+
+## 0.12.0
+
+### Patch Changes
+
+- @gtmgrid/auth@0.12.0
+
 ## 0.11.1
 
 ### Patch Changes